Configuration update for autobench
Expansion of possibilities for configuring tests. Added parameter validation for FIO.
Updated work with the catalog for collecting results.
Added the ability to change the target for the FIO load (for example, for a block device).
Fixed problem with restarting the test if there are old results.
And other minor bug fixes.