Update for multiple VM launches with testing.
Added general arguments for specifying the number of virtual machines, and the starting port for ssh connection.
Adapted multi-launch for VM number.
Thus, if the number of VMs is specified more than 1, in this case, the specified number of VMs will be created with the IP address specified in the test and ports starting from the specified
For example, if the initial port is 6666, then the next VM will have port 6667, and so on.